GM_01102018_General_1
Comentarios:
- Añadiendo sonido a los botones.
- Haciendo que el jugador no se mueva cuando ha muerto.
- Otros que no me acuerdo
