--- Assets/MenuMuerte.cs
@@ -1,18 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEditor.SceneManagement;
-
-public class MenuMuerte : MonoBehaviour {
-
-	// Use this for initialization
-	public void Retry () {
-        EditorSceneManager.LoadScene("Tutorial");
-	}
-    public void Menu()
-    {
-        EditorSceneManager.LoadScene("MENU");
-    }
-
-
-}
