+++ Assets/Scripts/GM/ClickSound.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 72180698e0c7a994a909649dca2fc48b
+timeCreated: 1515615683
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
