+++ Assets/Scripts/GM/MenuMuerte.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 875914b3866337b4a8ab4ee3ee6ac132
+timeCreated: 1515539019
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
