--- Assets/script1.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: fedb936ef46f6de49aee85fb885bbad0
-timeCreated: 1515535857
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
